% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_setup.R
\name{check_answers}
\alias{check_answers}
\title{Check learning task ranking accuracy}
\usage{
check_answers(data, col = NULL, bonus)
}
\arguments{
\item{data}{Untouched Gorilla task data. check.names=FALSE}

\item{col}{Vector of column names for first round cleaning. Already includes cols 'Response' 'display' and 'Participant Public ID'. Defaults to treehouse task names.}

\item{bonus}{double representing bonus payout per correct choice.}
}
\value{
Array of subjID that failed >2 questions, or no message else.
}
\description{
Function to check accuracy ranking.
}
