% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention_check_related.R
\name{check_attention}
\alias{check_attention}
\title{Check attention questions}
\usage{
check_attention(
  data,
  subj_col = NULL,
  questions = NULL,
  correct_answers = NULL
)
}
\arguments{
\item{data}{Untouched questionnaire data - short form with name only when downloading from Gorilla. check.names=FALSE.}

\item{subj_col}{Name of the column to be used to identify subjects. Defaults to "Participant Public ID".}

\item{questions}{Vector of exact question as written. Defaults to DAS lab 2025 version.}

\item{correct_answers}{Vector of correct answers in order of the questions. Defaults to DAS lab 2025 version.}
}
\value{
Array of subjID that failed >2 questions, or no message else.
}
\description{
Function to return participant id if failed more than 2 questions.
}
