% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_processing.R
\name{extract_questions}
\alias{extract_questions}
\title{Gather questionnaire data}
\usage{
extract_questions(data, subjID_col = NULL, remove = NULL, sum_only = TRUE)
}
\arguments{
\item{data}{raw questionnaire csv as exported per guideline.}

\item{subjID_col}{colname to represent unique subject ID. Defaults to Participant Public ID.}

\item{remove}{vector of any questions in between questionnaire items to remove. Defaults to DAS lab 2025 attention check questions.}

\item{sum_only}{do you want only the sum scores? Defaults to TRUE.}
}
\value{
Processed questionnaire data frame
}
\description{
Function to process raw questionnaire csv from Gorilla.
}
