% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorillatamer-package.R
\docType{package}
\name{gorillatamer-package}
\alias{gorillatamer}
\alias{gorillatamer-package}
\title{gorillatamer: Manages exported data from Gorilla.sc}
\description{
Tames the Gorilla.sc data.
}
\author{
\strong{Maintainer}: Napat Burapath \email{k.burapath@unsw.edu.au}

}
\keyword{internal}
