% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_setup.R
\name{process_raw_data}
\alias{process_raw_data}
\title{Check learning task ranking accuracy}
\usage{
process_raw_data(raw_task_data, cols = NULL)
}
\arguments{
\item{cols}{Vector of column names for first round cleaning. Defaults to treehouse task names.}

\item{data}{Untouched Gorilla task data. check.names=FALSE}
}
\value{
list(task, ranking) data frames.
}
\description{
Function to check accuracy ranking.
}
