% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_setup.R
\name{check_attention}
\alias{check_attention}
\title{Check attention questions}
\usage{
check_attention(data, questions = NULL, correct_answers = NULL)
}
\arguments{
\item{data}{Untouched questionnaire data - short form with name only when downloading from Gorilla. check.names=FALSE.}

\item{questions}{Vector of exact question as written. Defaults to DAS lab 2025 version.}

\item{correct_answers}{Vector of correct answers in order. Defaults to DAS lab 2025 version.}
}
\value{
Array of subjID that failed >2 questions, or no message else.
}
\description{
Function to return participant id if failed more than 2 questions.
}
